Add static analysis tests
