+++ tests/static-analysis/assert-count.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Webmozart\Assert\Tests\StaticAnalysis;
+
+use Countable;
+use Webmozart\Assert\Assert;
+
+/**
+ * @psalm-pure
+ * @param Countable|array $value
+ */
+function consume($value): void
+{
+    Assert::count($value, 0);
+}
