+++ tests/static-analysis/assert-regex.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Webmozart\Assert\Tests\StaticAnalysis;
+
+use Webmozart\Assert\Assert;
+
+/**
+ * @psalm-pure
+ */
+function consume(string $value): void
+{
+    Assert::regex($value, '/.+/');
+}
