+++ tests/static-analysis/assert-stringNotEmpty.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Webmozart\Assert\Tests\StaticAnalysis;
+
+use Webmozart\Assert\Assert;
+
+/**
+ * @psalm-pure
+ */
+function consume(string $value): void
+{
+    Assert::stringNotEmpty($value);
+}
